feat: add camera and start workout page
