+++ client/app.js
@@ -7,7 +7,6 @@ const App = () => {
   return (
     <div>
       <Navbar />
-
       <Routes />
     </div>
   )
