+++ public/index.html
@@ -3,12 +3,9 @@
   <head>
     <meta charset='utf-8'>
     <meta name='viewport' content='width=device-width initial-scale=1.0'>
-    <title>Boilermaker</title>
+    <title>LIFT LOG</title>
     <link rel="stylesheet" href="/style.css" />
     <script defer src="/bundle.js"></script>
-    <script src="https://cdn.jsdelivr.net/npm/@tensorflow/tfjs@1.3.1/dist/tf.min.js"></script>
-    <script src="https://cdn.jsdelivr.net/npm/@teachablemachine/pose@0.8/dist/teachablemachine-pose.min.js"></script>
-
   </head>
   <body>
     <div id="app"></div>
