+++ teachablemachine/metadata.json
@@ -1,0 +1,18 @@
+{
+  "tfjsVersion": "1.3.1",
+  "tmVersion": "0.8.6",
+  "packageVersion": "0.8.6",
+  "packageName": "@teachablemachine/pose",
+  "timeStamp": "2020-07-01T21:12:29.346Z",
+  "userMetadata": {},
+  "modelName": "Exercise Poses",
+  "labels": ["Neutral - Standing", "Bicep Curl - Up "],
+  "modelSettings": {
+    "posenet": {
+      "architecture": "MobileNetV1",
+      "outputStride": 16,
+      "inputResolution": 257,
+      "multiplier": 0.75
+    }
+  }
+}
