+++ teachablemachine/model.json
@@ -1,0 +1,88 @@
+{
+  "modelTopology": {
+    "class_name": "Sequential",
+    "config": {
+      "name": "sequential_2",
+      "layers": [
+        {
+          "class_name": "Dense",
+          "config": {
+            "units": 100,
+            "activation": "relu",
+            "use_bias": true,
+            "kernel_initializer": {
+              "class_name": "VarianceScaling",
+              "config": {
+                "scale": 1,
+                "mode": "fan_in",
+                "distribution": "normal",
+                "seed": null
+              }
+            },
+            "bias_initializer": {"class_name": "Zeros", "config": {}},
+            "kernel_regularizer": null,
+            "bias_regularizer": null,
+            "activity_regularizer": null,
+            "kernel_constraint": null,
+            "bias_constraint": null,
+            "name": "dense_Dense1",
+            "trainable": true,
+            "batch_input_shape": [null, 14739],
+            "dtype": "float32"
+          }
+        },
+        {
+          "class_name": "Dropout",
+          "config": {
+            "rate": 0.5,
+            "noise_shape": null,
+            "seed": null,
+            "name": "dropout_Dropout1",
+            "trainable": true
+          }
+        },
+        {
+          "class_name": "Dense",
+          "config": {
+            "units": 2,
+            "activation": "softmax",
+            "use_bias": false,
+            "kernel_initializer": {
+              "class_name": "VarianceScaling",
+              "config": {
+                "scale": 1,
+                "mode": "fan_in",
+                "distribution": "normal",
+                "seed": null
+              }
+            },
+            "bias_initializer": {"class_name": "Zeros", "config": {}},
+            "kernel_regularizer": null,
+            "bias_regularizer": null,
+            "activity_regularizer": null,
+            "kernel_constraint": null,
+            "bias_constraint": null,
+            "name": "dense_Dense2",
+            "trainable": true
+          }
+        }
+      ]
+    },
+    "keras_version": "tfjs-layers 1.3.1",
+    "backend": "tensor_flow.js"
+  },
+  "weightsManifest": [
+    {
+      "paths": ["weights.bin"],
+      "weights": [
+        {
+          "name": "dense_Dense1/kernel",
+          "shape": [14739, 100],
+          "dtype": "float32"
+        },
+        {"name": "dense_Dense1/bias", "shape": [100], "dtype": "float32"},
+        {"name": "dense_Dense2/kernel", "shape": [100, 2], "dtype": "float32"}
+      ]
+    }
+  ]
+}
